cb_uniswap_web_app_implementation/requirements.txt
pandas
numpy
requests
matplotlib
mplfinance
plotly
anthropic
web3
python-dotenv
flask

